move third party prints away
